Remove unneeded interfaces
