+++ app/src/main/java/com/wilderpereira/lmgtfygen/dagger/component/MainComponent.kt
@@ -1,7 +1,6 @@
 package com.wilderpereira.lmgtfygen.dagger.component
 
 import com.wilderpereira.lmgtfygen.dagger.module.NetworkModule
-import com.wilderpereira.lmgtfygen.dagger.module.PresenterModule
 import com.wilderpereira.lmgtfygen.presentation.MainActivity
 import dagger.Component
 import javax.inject.Singleton
@@ -10,7 +9,7 @@ import javax.inject.Singleton
  * Created by Wilder on 25/01/17.
  */
 @Singleton
-@Component(modules = arrayOf(NetworkModule::class, PresenterModule::class))
+@Component(modules = arrayOf(NetworkModule::class))
 interface MainComponent {
     fun inject(activity: MainActivity)
 }
